requirements.txt
anyio==4.2.0
certifi==2023.11.17
et-xmlfile==1.1.0
exceptiongroup==1.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.4
PyMySQL==1.1.0
python-dateutil==2.8.2
python-telegram-bot==20.7
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.24
typing_extensions==4.9.0
tzdata==2023.4
unicode==2.9
Unidecode==1.3.7